+++ include/beast/http.hpp
@@ -13,6 +13,7 @@
 #include <beast/http/basic_dynamic_body.hpp>
 #include <beast/http/basic_parser.hpp>
 #include <beast/http/buffer_body.hpp>
+#include <beast/http/chunk_encode.hpp>
 #include <beast/http/dynamic_body.hpp>
 #include <beast/http/empty_body.hpp>
 #include <beast/http/error.hpp>